% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWR.R
\name{NWR}
\alias{NWR}
\title{NWR -  A function to run network weighted regression models by performing optimization of the posterior distribution.}
\usage{
NWR(
  formula,
  data,
  y = FALSE,
  X = FALSE,
  A,
  k = 5,
  gamma_prior = list(mean = 0, sigma = 1),
  beta0_prior = list(mean = 1, sd = 1),
  s2_prior = list(shape = 1/2, scale = 1/2),
  rho_prior = list(mean = 0.36, sd = 0.7^2),
  ...
)
}
\arguments{
\item{formula}{A formula line usually denoted y ~ X or outcome ~ predictors.}

\item{data}{The dataset the variables from the formula line come from.}

\item{y}{If you would rather input an outcome vector directly. Must be used in conjunction with X and without a formula.}

\item{X}{If you would rather input a predictor matrix directly. DO NOT include an intercept column, the intercept is modeled differently from the other predictors. Must be used in conjunction with y and without a formula.}

\item{A}{An adjacency matrix describing your network. It will be row-normalized in the function if it is not already.}

\item{k}{The number of important eigenvectors of A. The true value of this number is probably unknown to you but you can test for optimal k using the model selection method of your choice and the NWR_kstep function. The larger your k, the longer the function will take to run (the more parameters there will need to be estimated). A good place to start might be the number of predictors.}

\item{gamma_prior}{The prior for each gamma vector of the model. Requires list or function inputs. Currently requires all vectors to have the same prior. Default is a multivariate normal with mean=0 and sigma=I where I is the identity matrix. If you supply a list, you must have one value named 'mean' which is the mean of the multivariate normal and one named 'sigma' which is the variance-covariance matrix. If mean is a single integer, that mean will be used for every element in the mean vector and if sigma is a single integer, it will multiply that integer by the identity matrix. Otherwise, both mean and sigma will be what you specify them to be. Keep in mind, if you're adjacency matrix is not symmetrical (when it isn't row-normalized), you will need the vectors to be 2(k+1)p in dimension. And if it is symmetrical, they need to be (k+1)p in dimension. You can also supply your own prior by supplying a function name of the prior function you would like to use.}

\item{beta0_prior}{The prior for the intercept of the model. Requires list or function inputs. Default is a N(0,1). If you supply a list, you must have one value named 'mean' which is the mean of the normal and one named 'sd' which is the standard deviation with both values of length=1. You can also supply your own prior by supplying a function name of the prior function you would like to use.}

\item{s2_prior}{The prior for the variance of the model. Requires list or function inputs. Default is an inverse-gamma(0.5,0.5). If you supply a list, you must have one value named 'shape' which is the shape of the inverse-gamma and one named 'scale' which is the scale with both values of length=1. You can also supply your own prior by supplying a function name of the prior function you would like to use.}

\item{rho_prior}{The prior for the correlation of the adjacency matrix of the model. Requires list or function inputs. Default is a N(0.36,0.49) (Dittrich et al. 2017). If you supply a list, you must have one value named 'mean' which is the mean of the normal and one named 'sd' which is the standard deviation with both values of length=1. You can also supply your own prior by supplying a function name of the prior function you would like to use.}

\item{...}{Values to be passed into the optim() function. Currently defaults to: optim(par=c(rep(0, length(gamma_mat)),1,0.2), fn=posterior_function, gr=NULL, method="L-BFGS-B", lower=c(rep(-15, length(gamma_mat)), 1e-3, -.85), upper=c(rep(15, length(gamma_mat)), 20, .85), control=list(fnscale=-1), hessian=TRUE). I don't allow you to change fn or hessian.}
}
\value{
A list containing values:

\code{beta0} The estimated intercept of the model.

\code{beta} The estimated coefficient matrix of the model.

\code{s2} The estimated variance of the model.

\code{rho} The estimated correlation of the adjacency matrix of the model.

\code{gamma} The estimated gamma parameters of the model.

\code{U} The Orthonormal matrix of k important eigenvectors of the spectral or QR decomposition of the model (Plus intercept column).

\code{A} The inputted adjacency matrix (row-normalized).

\code{H} The hessian matrix of the optimization.

\code{Hinv} The inverse of the hessian matrix.

\code{logpost} The log of the posterior density.

\code{y} The inputted outcome vector.

\code{X} The inputted prediction matrix.
}
\description{
NWR -  A function to run network weighted regression models by performing optimization of the posterior distribution.
}
