% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.NWR.R
\name{plot.NWR}
\alias{plot.NWR}
\title{plot.NWR -  A function that plots the estimated coefficient vector(s) of an NWR object.}
\usage{
\method{plot}{NWR}(NWR, seed = NULL, boxplot_args, igraph_args)
}
\arguments{
\item{NWR}{An NWR object.}

\item{seed}{The random seed to set if you want the plot to change orientation every time you run the function.}

\item{boxplot_args}{Additional arguments to go into the boxplot() function.}

\item{igraph_args}{Additional arguments to go into the igraph() funciton.}
}
\value{
Boxplot(s) of the distribution of the coefficient matrix and network plots for each coefficient vector.
}
\description{
plot.NWR -  A function that plots the estimated coefficient vector(s) of an NWR object.
}
