% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.NWR.R
\name{summary.NWR}
\alias{summary.NWR}
\title{summary.NWR -  A function that summarizes the results of an NWR object.}
\usage{
\method{summary}{NWR}(NWR, ci = 0.95)
}
\arguments{
\item{NWR}{An NWR object.}

\item{ci}{The level of the credible interval.}
}
\value{
A list with values:

\code{Parameters} The estimated parameters (with mean of each coefficient vector) as well as their 100*ci\% credible intervals.

\code{logpost} The log of the posterior of the estimated parameters.

\code{BIC} The BIC of the model.

\code{AICc} The corrected AIC of the model. Corrected AIC is suggested to be used when comparing the NWR model to other models in order to account for the bias of AIC which doesn't penalize models with too many parameters as much as it should. Doing so is important for a model with a matrix of parameters of unspecified size.
}
\description{
summary.NWR -  A function that summarizes the results of an NWR object.
}
